complete 2 — Database Access Using the same provided basic web framework, and data fixtures for authors and posts that are provided:
Write an importer in PHP that imports a list of post files (examples are provided in the data folder) into the database.
Update src/Controller/PostDetails.php to load a published post from the database with the specified id. Update the src/Template/PostDetails.php template to render the post content (title, body, author) as HTML. The post body is formatted as Markdown and the HTML should include the formatted Markdown.
Update the src/Controller/PostIndex.php to load all published posts from the database in reverse chronological order. Update the src/Template/PostIndex.php template to render ths list as HTML. Include the post titles and authors in the output. Make clicking a post go to the post details.